[move-cli] A proper "docgen" implementation (#327)
A rudimentary version of the "docgen" command was implemented in
55a40d583 ("[move-cli] Added "docgen" command (#291)", 2022-07-28).
However, the implementation was relying on the Move Prover, only
exposing one Docgen switch to the user.

This implementation is using the actual Dogen, giving us a possibility
to expose all the Docgen switches to the user. The enable/disable
logic is modeled against current default settings for Docgen. This
is also an indirect way to communicate the default settings to the
user.

Addressing the actual Docgen as move_docgen::Docgen, because
both, this subcommand and the Move Docgen are entitled to that name:
this subcommand because of the Move CLI coding convention, and the
actual Move prover because it's the most descriptive short name.

Removed --quiet, since output on successful execution is now
deterministic.

The test is currently testing all the switches, creating (almost)
totally opposite settings from the default settings.